pyproject.toml
[project]
name = "olmo-core_snellius"
version = "0.1.0"
description = "OLMo core ported to Snellius"
authors = [{ name = "tvosch", email = "thomas.vanosch@surf.nl" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
    "ai2-olmo-core",
    "wandb",         
]

[dependency-groups]
flash-attention = [
    "flash-attn"
]